requirements.txt
Django==4.1.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
